+++ api/src/routes/auth.route.ts
@@ -1,16 +1,24 @@
 import { Router } from "express";
 import authController from "../controllers/auth.controller";
 import authenticateToken from "../middleware/authenticateToken";
+import validateParams, { roleSchema } from "../validators/params.validator";
+import authenticateRole from "../middleware/authenticateRole";
 
 const authRoutes = () => {
-  const { authorize, authenticatedUser, refreshToken, login, logout, signup } =
-    authController;
+  const {
+    authorize,
+    getAuthenticatedUser,
+    refreshToken,
+    login,
+    logout,
+    signup,
+  } = authController;
   const authRouter = Router();
 
   // GET requests
   authRouter.get("/", authenticateToken, authorize);
-  authRouter.get("/user", authenticateToken, authenticatedUser);
-  authRouter.get("/user/:username", authenticateToken, authenticatedUser);
+  authRouter.get("/user", authenticateToken, getAuthenticatedUser);
+  authRouter.get("/user/:username", authenticateToken, getAuthenticatedUser);
 
   // POST requests
   // curl -w "\n" -X POST http://localhost:3001/api/auth